requirements.txt
# =============================================================================
# APEX — Auto-Trading AI Agent
# requirements.txt — Python Dependencies
# =============================================================================
# Install on server with:
# pip3 install -r requirements.txt --user
# =============================================================================

# Exchange connectivity
ccxt==4.2.15

# Data processing
pandas==2.0.3
numpy==1.24.4

# Technical indicators
ta==0.11.0

# HTTP requests (Fear & Greed, CoinGecko, Telegram)
requests==2.31.0

# Scheduling
APScheduler==3.10.4

# Dashboard
streamlit==1.29.0
plotly==5.18.0

# Database
# sqlite3 is built into Python — no install needed

# Dashboard styling
jinja2==3.1.2

# Environment variables
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
